fix an mdi queueing bug
This commit fixes a bug with MDI queueing in Task.  MDI queueing is used
when a UI sends MDI commands to Task, but the interpreter is busy running
previous commands.  MDI commands in this case are read from the NML
Command channel and placed into a queue inside Task.  When the interpreter
finishes its current job, Task feeds it the next command from the queue.

The bug came from a confusion inside Task about whether we're running old
commands from the MDI queue or new commands from the NML Command channel.
This commit unconfuses Task.