rtapi/xenomai: fix race condition on RT thread deletion
An rt_task_delete() does not make the thread go away immediately.

To assure the call returns only after the thread is actually terminated,
create threads with T_JOINABLE and rt_task_join() the thread on
RTAPI thread deletion.

this caused segfaults in RTAPI on shutdown, where the rtapi.so module
was already unloaded but the RT thread was not finished

Incidentially, this makes the USE_SIGXCPU support make work reliable.
I'll leave it turned off though - signal handlers in RT threads are
potentially fishy.

TODO: this fix should be ported to xenomai-kernel too