Undo hack to overrun test
This commit broke the overrun test:  05c8f859

This hack fixed it:  e3237db6

After the 'autoload' feature, 05c8f859 was reversed:  164647d4

Now the overrun test is breaking on the buildbots, even though not on
my various test environments.

This commit reverses 05c8 because it's right, and hopefully the
buildbots will stop getting hosed on this test.